requirements-whois.txt
fastapi
uvicorn[standard]
python-whois
